去掉从Game.Scene直接Get Scene,应该使用domain来获得
