+++ Server/Hotfix/Scene/SceneFactory.cs
@@ -32,7 +32,7 @@ namespace ET
                     break;
                 case SceneType.Map:
                     scene.AddComponent<UnitComponent>();
-                    scene.AddComponent<RecastPathComponent>();
+                    //scene.AddComponent<RecastPathComponent>();
                     break;
                 case SceneType.Location:
                     scene.AddComponent<LocationComponent>();
