--- Server/Hotfix/Scene/SceneHelper.cs
@@ -1,6 +1,0 @@
-namespace ET
-{
-    public static class SceneHelper
-    {
-    }
-}
