+++ Unity/Assets/Hotfix/Unit/UnitFactory.cs
@@ -14,7 +14,7 @@ namespace ET
 
 	        Game.EventSystem.Publish(new EventType.AfterUnitCreate() {Unit = unit});
 	        
-	        UnitComponent unitComponent = Game.Scene.Get(1).GetComponent<UnitComponent>();
+	        UnitComponent unitComponent = domain.GetComponent<UnitComponent>();
             unitComponent.Add(unit);
             return unit;
         }
