+++ Unity/Assets/HotfixView/Opera/OperaComponentSystem.cs
@@ -33,7 +33,7 @@ namespace ET
                     self.frameClickMap.X = self.ClickPoint.x;
                     self.frameClickMap.Y = self.ClickPoint.y;
                     self.frameClickMap.Z = self.ClickPoint.z;
-                    self.ZoneScene().GetComponent<SessionComponent>().Session.Send(self.frameClickMap);
+                    self.DomainScene().GetComponent<SessionComponent>().Session.Send(self.frameClickMap);
 
                     // 测试actor rpc消息
                     self.TestActor().Coroutine();
@@ -45,7 +45,7 @@ namespace ET
         {
             try
             {
-                M2C_TestActorResponse response = (M2C_TestActorResponse)await self.ZoneScene().GetComponent<SessionComponent>().Session.Call(
+                M2C_TestActorResponse response = (M2C_TestActorResponse)await self.Domain.GetComponent<SessionComponent>().Session.Call(
                     new C2M_TestActorRequest() { Info = "actor rpc request" });
                 Log.Info(response.Info);
             }
