--- Unity/Assets/HotfixView/Scene/SceneHelper.cs
@@ -1,10 +1,0 @@
-namespace ET
-{
-    public static class SceneHelper
-    {
-        public static Scene ZoneScene(this Entity entity)
-        {
-            return Game.Scene.Get(entity.DomainZone());
-        }
-    }
-}
