+++ Unity/Assets/Model/Core/Entity/Scene.cs
@@ -95,17 +95,4 @@
             }
         }
     }
-
-    public static class SceneEx
-    {
-        public static int DomainZone(this Entity entity)
-        {
-            return ((Scene) entity.Domain)?.Zone ?? 0;
-        }
-
-        public static Scene DomainScene(this Entity entity)
-        {
-            return (Scene) entity.Domain;
-        }
-    }
 }
