+++ Unity/Assets/Model/Scene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: da44f331d6f7c4109be8c2a5f7672697
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
